Add embeds to message command and other stuff

+++ commands/say.js
@@ -1,10 +1,11 @@
 module.exports = {
     name: 'say',
-    description: 'Repeats what you say.',
+    description: 'Repeats what you say',
     category: 'fun',
+    usage: '[message]',
     args: true,
     execute(message, args) {
-        message.delete();
+        if (!message.channel.type === 'dm') message.delete();
         const sayText = args.join(' ');
 
         const channel = message.client.channels.cache.find(channel => channel.id === '749084221024239717');
